Refactor the websockets server code
Mostly just moves things around. Coming up is the addition of a database
cloner (so we have all available lines and are able to query for all
SPTrans buses' positions). Once this is done, the server will be worked
on in order to be published as a more convenient way to access SPTrans'
APIs.