Start writting a mock server for testing
And fun&profit. This is good for me as a Haskell developer, because it
seems like the standard practice in Haskell is to generally to build the
kind of impure tests one would write closer to metal on other more
flexible languages at the E2E level - I need to be able to do this
reasonably fast.

And I just like writting all sorts of Haskell, so I don't mind the extra
work, honestly.