Fix and add tests to `queryStops`
It's also a little documented, though I'm not sure haddock is smart
enough to do this.

It uses scoped type variables to use a constant function that sort of
takes a phantom type, in order to hold a constant in the `StopsQuery`
typeclass. It's a little silly, but it's a decent payoff to have the
nicer interface, cutting off two functions in API area.

Closes #1